pyproject.toml
[tool.poetry]
name = "data_pipeline"
version = "0.1.0"
description = "A simple data pipeline to extract data from sql database and send it to a bucket"
authors = ["Rob Bailiff <rob.bailiff@airelogic.com>"]

[tool.poetry.dependencies]
python = "^3.8"
minio = "^7.1.2"
pandas = "^1.3.5"
SQLAlchemy = "^1.4.29"
psycopg2-binary = "^2.9.3"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
boto3 = "^1.21.22"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
